fragments moved to UI
