+++ android/app/src/main/java/com/team4infinity/smartyfridge/ui/fragments/FridgeFragment.kt
@@ -1,4 +1,4 @@
-package com.team4infinity.smartyfridge.fragments
+package com.team4infinity.smartyfridge.ui.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -14,10 +14,10 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [RecipesFragment.newInstance] factory method to
+ * Use the [FridgeFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class RecipesFragment : Fragment() {
+class FridgeFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -35,7 +35,7 @@ class RecipesFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_recipes, container, false)
+        return inflater.inflate(R.layout.fragment_fridge, container, false)
     }
 
     companion object {
@@ -45,12 +45,12 @@ class RecipesFragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment RecipesFragment.
+         * @return A new instance of fragment FridgeFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            RecipesFragment().apply {
+            FridgeFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
