+++ android/app/src/main/res/layout/fragment_recipes.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.RecipesFragment">
+    tools:context=".ui.fragments.RecipesFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
