+++ android/app/src/main/res/navigation/main_nav.xml
@@ -7,17 +7,17 @@
 
     <fragment
         android:id="@+id/bnb_home"
-        android:name="com.team4infinity.smartyfridge.fragments.MainFragment"
+        android:name="com.team4infinity.smartyfridge.ui.fragments.MainFragment"
         android:label="fragment_main"
         tools:layout="@layout/fragment_main" />
     <fragment
         android:id="@+id/bnb_fridge"
-        android:name="com.team4infinity.smartyfridge.fragments.FridgeFragment"
+        android:name="com.team4infinity.smartyfridge.ui.fragments.FridgeFragment"
         android:label="fragment_fridge"
         tools:layout="@layout/fragment_fridge" />
     <fragment
         android:id="@+id/bnb_recipes"
-        android:name="com.team4infinity.smartyfridge.fragments.RecipesFragment"
+        android:name="com.team4infinity.smartyfridge.ui.fragments.RecipesFragment"
         android:label="fragment_recipes"
         tools:layout="@layout/fragment_recipes" />
 </navigation>
